requirements.txt
streamlit==1.24.0
pandas==1.5.3
numpy==1.24.3
pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
plotly==5.15.0
volcenginesdkarkruntime==1.0.0